add patient controller and routes
