+++ appointment-svc/middlewares/hash-middleware.ts
@@ -1,0 +1,5 @@
+import { NextFunction } from 'express';
+
+const hashMiddlware = (req: Request, res: Response, next: NextFunction) => {};
+
+export default hashMiddlware;
